msm: copper: Add target specific dts files
The filename msmcopper.dts does not describe a target, split
these up to create SoC specific (.dtsi) and board specific
(.dts) device tree source files. The board specific files
need to include the SoC file and can selectively specify the
devices that are not present or supported.

Update the makefile to generate all board variants of appended
DTB zImage files or allow for a specific variant to be generated
using the DTS_NAME override.

Change-Id: Ic0bba8313866352dde67287fab88aa20b77738f0
Signed-off-by: Sathish Ambley <sambley@codeaurora.org>

Conflicts:
	AndroidKernel.mk
	arch/arm/boot/dts/msmcopper.dtsi