msm: bam_dmux: add missing parameter to bam_dmux_log() call
bam_dmux_log() takes a formatted string and a variable argument list
similar to printf() and creates printf() style output to a private kfifo
log.  This change fixes a particular invocation of bam_dmux_log() which has
an invalid variable argument list for the formatted string used.  Depending
on system timing, this invocation could succeed as expected, or it could
cause a null pointer exception.

To prevent further issues, add the __printf() function attribute so that
the gcc compiler will check at compile time the arguments to bam_dmux_log()
for proper number and types according to printf() usage.

CRs-Fixed: 382037
Change-Id: I247ac3e83664385b9dc61a434049ca5c8bb60cd2
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>