mmc: msm_sdcc: Check for unclocked register access in IRQ handler
After SDCC has been suspended and clocks turned off, an asynchronous
interrupt from an SDIO device will result in a call to the IRQ handler.
Since SDCC hardware registers are unclocked, a read of the status
register will result in an exception. Verify that clock to SDCC is on
before access in the IRQ handler.

CRs-Fixed: 424960
Change-Id: I81afd2dd5c5fa4785f11f2eec03e5b454b8f5dbe
Signed-off-by: Oluwafemi Adeyemi <aadeyemi@codeaurora.org>