mmc: msm_sdcc: acquire wakelock only if needed
Currently the driver assumes that when it gets an interrupt with clocks
turned OFF, that it has gotten a sdio wake up interrupt and tries to take
a wakelock. While this is not expected when sdio interrupt pin or sdio
interrupt is not defined, it looks like a situation we run into sometimes.
Ensure that the interrupt handler takes the wakelock only if the sdio
interupt or sdio interrupt pin is defined.

CRs-Fixed: 454585
Change-Id: I4842e293a5ce19a1c84a487f836f23001d69edac
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>