M2: pm: Remove ccadc calibration schedule
As of
99ee5 msm: power: update power drivers to msm-3.4 tip
ccadc re-calibration is forced for all pm8921 users, both on resume, and
on during regular time-intervals via a delayed workqueue.

Locking a mutex in the resume path
that is also locked in a workqueue that is allowed to sleep is poor
design.  In particular, if the system suspends while the
mutex is held, trying to re-lock the mutex in the resume path causes
deadlock (SOD).

Could be fixed with mutex_trylock in the resume path, but:

This is not for us, since in the stock mangled samsung pm8xxx drivers, there is no
ccadc (re)calibration. Not calibrating seems to not affect battery life.

Change-Id: Icbc9c1c494417e4b718106a49b2da24ec093c4e2