msm: pm-8x60: Remove timeout waiting for cpu to power collapse
The delays in the CPU_DYING notication path could result in the core initiating the hotplug to send out CPU_DEAD notifications after msm_pm_wait_cpu_shutdown() times out. This could result in a unknown resets when regulator/clocks are disabled in the CPU_DEAD path. To prevent unknown resets, wait infinitely to enter power collapse but throw a warning after the 1ms mark.

The platform_cpu_kill() path is also executed in the reboot path when a core tries to stop secondary cores. In these cases, platform_cpu_kill shouldn't wait infinitely for the secondary cores to power collapse. To circumvent this scenario, track dying cpus as a part of the hotplug notification and skip waiting for the core to power collpase during reboot path.

Change-Id: I144e95a6d019e897381d2bca62ab50ecefd56b5b
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>